Cleaned up code automatically

+++ src/main/java/net/journey/entity/mob/frozen/EntityShatterer.java
@@ -1,11 +1,8 @@
 package net.journey.entity.mob.frozen;
 
-import java.util.Random;
-
 import net.journey.JourneyBlocks;
 import net.journey.JourneySounds;
 import net.journey.entity.MobStats;
-import net.journey.entity.mob.terrania.mob.EntityTerralight;
 import net.minecraft.entity.ai.EntityAIBase;
 import net.minecraft.entity.ai.EntityAIFindEntityNearestPlayer;
 import net.minecraft.entity.ai.EntityMoveHelper;
@@ -18,61 +15,64 @@ import net.minecraft.world.EnumDifficulty;
 import net.minecraft.world.World;
 import net.slayer.api.entity.EntityModFlying;
 
+import java.util.Random;
+
 public class EntityShatterer extends EntityModFlying {
 
-	public EntityShatterer(World par1World) {
-		super(par1World);
-		this.moveHelper = new EntityShatterer.ShattererMoveHelper();
+    public EntityShatterer(World par1World) {
+        super(par1World);
+        this.moveHelper = new EntityShatterer.ShattererMoveHelper();
         initEntityAI();
-		setSize(2F, 2F);
-	}
-	public void initEntityAI()
-    {
+        setSize(2F, 2F);
+    }
+
+    public void initEntityAI() {
         this.tasks.addTask(5, new EntityShatterer.AIRandomFly());
         this.targetTasks.addTask(1, new EntityAIFindEntityNearestPlayer(this));
     }
-	@Override
-	public void onUpdate() {
+
+    @Override
+    public void onUpdate() {
         super.onUpdate();
-        if(!this.world.isRemote && this.world.getDifficulty() == EnumDifficulty.PEACEFUL) this.setDead();
+        if (!this.world.isRemote && this.world.getDifficulty() == EnumDifficulty.PEACEFUL) this.setDead();
     }
 
-	@Override
-	public double setMaxHealth(MobStats s) {
-		return MobStats.ShattererHealth;
-	}
+    @Override
+    public double setMaxHealth(MobStats s) {
+        return MobStats.ShattererHealth;
+    }
 
-	@Override
-	public SoundEvent setLivingSound() {
-		return JourneySounds.HONGO;
-	}
+    @Override
+    public SoundEvent setLivingSound() {
+        return JourneySounds.HONGO;
+    }
 
-	@Override
-	public SoundEvent setHurtSound() {
-		return JourneySounds.SAND_CRAWLER;
-	}
+    @Override
+    public SoundEvent setHurtSound() {
+        return JourneySounds.SAND_CRAWLER;
+    }
 
-	@Override
-	public SoundEvent setDeathSound() {
-		return JourneySounds.SAND_CRAWLER;
-	}
-	
-	@Override
-	public Item getItemDropped() {
-		return null;
-	}
-	
-	@Override
-	public boolean shouldRenderInPass(int pass) {
-		return pass == 1;
-	}
-	
-	@Override
-	public boolean getCanSpawnHere() {
-		return this.world.getBlockState(new BlockPos(this.posX, this.posY-1, this.posZ)).getBlock() == JourneyBlocks.brittleIce;
-	}
-	
-	private class AIRandomFly extends EntityAIBase {
+    @Override
+    public SoundEvent setDeathSound() {
+        return JourneySounds.SAND_CRAWLER;
+    }
+
+    @Override
+    public Item getItemDropped() {
+        return null;
+    }
+
+    @Override
+    public boolean shouldRenderInPass(int pass) {
+        return pass == 1;
+    }
+
+    @Override
+    public boolean getCanSpawnHere() {
+        return this.world.getBlockState(new BlockPos(this.posX, this.posY - 1, this.posZ)).getBlock() == JourneyBlocks.brittleIce;
+    }
+
+    private class AIRandomFly extends EntityAIBase {
         private EntityShatterer e = EntityShatterer.this;
 
         public AIRandomFly() {
@@ -82,12 +82,12 @@ public class EntityShatterer extends EntityModFlying {
         @Override
         public boolean shouldExecute() {
             EntityMoveHelper entitymovehelper = this.e.getMoveHelper();
-            if(!entitymovehelper.isUpdating()) {
+            if (!entitymovehelper.isUpdating()) {
                 return true;
             } else {
-				double d0 = entitymovehelper.getX() - this.e.posX;
-				double d1 = entitymovehelper.getY() - this.e.posY;
-				double d2 = entitymovehelper.getZ() - this.e.posZ;
+                double d0 = entitymovehelper.getX() - this.e.posX;
+                double d1 = entitymovehelper.getY() - this.e.posY;
+                double d2 = entitymovehelper.getZ() - this.e.posZ;
                 double d3 = d0 * d0 + d1 * d1 + d2 * d2;
                 return d3 < 1.0D || d3 > 3600.0D;
             }
@@ -123,28 +123,28 @@ public class EntityShatterer extends EntityModFlying {
                 double d1 = this.posY - this.e.posY;
                 double d2 = this.posZ - this.e.posZ;
                 double d3 = d0 * d0 + d1 * d1 + d2 * d2;
-                if(this.height-- <= 0) {
+                if (this.height-- <= 0) {
                     this.height += this.e.getRNG().nextInt(5) + 2;
-                    d3 = (double)MathHelper.sqrt(d3);
-                    if(this.canMove(this.posX, this.posY, this.posZ, d3)) {
+                    d3 = MathHelper.sqrt(d3);
+                    if (this.canMove(this.posX, this.posY, this.posZ, d3)) {
                         this.e.motionX += d0 / d3 * 0.1D;
                         this.e.motionY += d1 / d3 * 0.1D;
                         this.e.motionZ += d2 / d3 * 0.1D;
                     } else {
-                    	this.action = EntityMoveHelper.Action.WAIT;
+                        this.action = EntityMoveHelper.Action.WAIT;
                     }
                 }
             }
         }
 
-        private boolean canMove(double x, double y, double z, double h)  {
+        private boolean canMove(double x, double y, double z, double h) {
             double d4 = (x - this.e.posX) / h;
             double d5 = (y - this.e.posY) / h;
             double d6 = (z - this.e.posZ) / h;
             AxisAlignedBB axisalignedbb = this.e.getEntityBoundingBox();
-            for(int i = 1; i < h; ++i) {
+            for (int i = 1; i < h; ++i) {
                 axisalignedbb = axisalignedbb.offset(d4, d5, d6);
-                if(!this.e.world.getCollisionBoxes(this.e, axisalignedbb).isEmpty()) {
+                if (!this.e.world.getCollisionBoxes(this.e, axisalignedbb).isEmpty()) {
                     return false;
                 }
             }
